ml-service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.25.2
requests>=2.31.0
python-dotenv>=1.0.0
geopy>=2.4.1
joblib>=1.3.2
python-multipart>=0.0.6
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1